Add error handling
Added error handling to
deletequestionnaire
doanswer
getallquestionnaires
getallsessions
getquestion
getquestionanswers
getquestionnaire
getsessionanswers
resetall
resetq

+++ api-backend/modules/resetq.js
@@ -4,9 +4,12 @@ const Answerraire = require('../models/answerraire')
 const Answer = require("../models/answer")
 
 const resetDB = async ({ questionnaireID }) => {
+    const questionnaire = await Questionnaire.findOne({ questionnaireID })
+    if (!questionnaire) throw ({ statusCode: 400, message: 'Questionnaire does not exist.' })
+
     await Answerraire.deleteMany({ questionnaireID })
     await Answer.deleteMany({ questionnaireID })
-    return 200
+    return { "status": "OK" }
 }
 
 module.exports = resetDB
